Major changes to the monitor abstraction.
Primarily, allows monitors to support multiple monitored objects of
different types; also defines a typed messaging medium.

This will support sending proxied XML SOAP requests from a server to
the analysis OR to its target(s) (that it created after the server launched
it).  Lots of i/o stuff, and locking, is bascially all this is.

Also, new test cases in tests/.