Be more deliberate about catching problems in waitpipe_add.
Sometimes, a process we launch might disappear before we can add it to
our waitpipe, so it won't get waited for because we don't track the sigchld.
I thought I fixed this in a previous commit, but apparently I was too
aggressive.

If we are ptrace'ing a process, and it spawns a new thread, and we try to
attach to that, we cannot call waitid() on that pid until we have been made
the ptrace'ing parent by the kernel.  But the process might exist.  So, if
we cannot wait on the process, check one more time for its existence via
kill(); if kill returns ESRCH, then the process really doesn't exist.  BUT
if the process exists by kill, allow the pid to be added.

This is not foolproof either, it's more a ptrace hack; there are cases for
more general uses of the library where the pid being added will *never* be
wait()able by us.  Let's call that user error and punt :).