Add a ResourceSchemaGenerator that attempts to read an existing schema.
Instead of letting axis introspect on the Service bean to generate XML
serializers, we try to use existing serializers and their XML schema
objects (which were already generated by axis, so should be safe!) by
getting their types from the WSDL file that axis read in the first place
when it generated the serializing java classes.  Then, just use the classes.

All this trouble is coming because we use xsd:token types -- somehow our
enumerations are restrictions of xsd:token, that's probably why it's
getting pulled in -- and axis doesn't seem to handle xsd:tokens correctly
during introspection.

This is hopefully ok to do because axis generated the XML stubs -- we're
just pulling them back in :).  We'll see.

This isn't fully working yet... hopefully it will.  Now at least we can
grab the schemas from the wsdl in a way axis will grok.  So much trouble...
tried wsdl4j but it doesn't give you access to the DOM inside a wsdl;
tried Woden but it only supports WSDL 2.0; gSOAP only seems to generate
WSDL 1.x... sigh.