Make an Axis-based service work from an existing schema and objects!
(pre-NB-NB: do not change the VMI XML namespace nor the WSDL location
URLs without understanding this commit fully!)

(pre-NB: I had to change the default WSDL ports from the simple
http://localhost:<port> form to
http://localhost:<port>/vmi/1/<servicename>!  See explanation near
bottom of this commit.  Oh, I also had to change the targetListener
interface slightly so that the method names were not the same as the
param names.  Well, I'm not sure I *had* to, but it would have been
confusing as ResourceSchemaGenerator.java tried to map the XML schema
types to existing axis-generated stub objects.)

NB: this commit message is important for posterity.

With Axis 1.6.2, it was impossible to pass only the service object
(descending from the service skeleton), and get things to work.  Why?
Axis is setup to autogenerate internal XML schema info that is inferred
from the service object's methods (which represent SOAP RPCs).  BUT, I
found it barfed when it encountered the XSD 'token' type (basic XML
stuff!); so it could not generate schema, wouldn't build the service.

So, I thought, ok, I have perfectly good XML schema; I just need to get
it read into Axis and tell Axis the SOAP-relevant XML types, and bind
them to the service object types and operations.  Ok, this was hard.

First, the XML schema info is buried in the WSDL XML file.  There wasn't
any easily available code to read WSDL 1.x, which is the thing gSOAP
seems to generate; only code to read WSDL 2.x.  Ok, so after realizing
this, I hacked some crappy DOM code to read the element containing
XML schema info out of either WSDL 1.x or 2.0 (I hope -- it's not
robust but all we need to process is *our* dang WSDLs -- so who cares).
Ok, so now I have the XML schema; then I have to basically do what
Axis does when building a service for a service object, BUT without
generating the XML schema during its processing.  And I have to
rebuild some of the internal state it needs.  Ok, that junk is all
done in ResourceSchemaGenerator.java.

Then there is some code to properly build a simple HTTP server and
configure it correctly, using an existing axis simple HTTP server.

The main things to note:

  * If the Axis people ever change their internal SchemaGenerator
    API, we'll probably have to sort of duplicate their work.

  * Not only must the service object you want to serve via SOAP RPC
    extend an axis-generated skeleton; it must implement the
    SimpleService interface, so we can get the info we need.
    See SimpleTargetListener.java .

  * Ah yes, the request processing path.  Even the javadoc has
    nothing in many places.  I read the code.

(The Axis request processing path is complex; it will make you want
to die as you debug it.  The only way I made it through is because
I used a debugger to step through and figure out why the paths I was
using weren't working by looking at values, and because they output
lots of debug output in the path.  Well, they would have to, or they
would go nuts!)

So, basically, the successful way I found to get requests through the
path required 1) WSDL service locations to be substrings of the target
XML namespace.  For us, the namespace is

  http://anathema.flux.utah.edu/schema/vmi/1

and now the service locations look like

  http://localhost:<port>/vmi/1/<servicename>.

This seems to be just enough to "fool" axis into doing something that
works.  It is able to extract <servicename>, demux it to a service,
then demux the service operation by reading the SOAP message.  It has
at least one other way that it tries to find the servicename and
operation name, looking for URLs like

  /prefix/<servicename>?<...>/<operationname>

or some crazy crap like that.  This path was "ensnaring" us until
I changed the WSDL locations for our services to be substrings of the
XML namespace.

SO: don't change the namespace nor the WSDL locations without
considering this fact!