Factor out the generic RPC stuff from target_rpc libs.
Then we can reuse more of it to build the analysis server more cleanly.

This cleans up locking, by passing the monitor a per-objtype mutex when
the objtype is registered.  Now when we lookup a monitored object id
(i.e., demux an incoming RPC on a target or analysis id), we can
optionally and atomically lock the objtype's mutex too.  This makes the
locking code cleaner and more reuseable; usually, when teh server looks
up an objid, it is going to need to hold its global lock -- and since the
servers are per-type -- using the per-type lock for the global server lock
makes good sense.

Of course, the monitor code NEVER unilaterally holds the per-objtype
mutexes; only when the user calls the lookup_and_lock functions will
it get locked.