Change the API for Listeners a bit.
Key outcomes: when you instantiate a target, you can specify an "owning"
listener.  Only owning listeners can answer Listener Notification messages
with responses that modify the state of the Probe or Action (i.e., disable
it, or remove it, or something).  Non-owning (non-authoritative) listeners
cannot modify such state.

Also, for owning listeners, I increased the connect, recv, and send timeouts
to a full day.  For non-owning listeners, it's 4 seconds each (those
notifications are best-effort and cannot block the server).

I didn't get as far as I wanted, so listeners can only bind/unbind to
Targets (not sub-objects like Probes or Actions -- that was really hard
and I wasted lots of time figuring out the model, and had to bail).

But, still an improvement.