Sigh, switch back over to implicit-register listener bind.
Had to do this because if the master server registers a listener and gives
it an ID, if that listener is later bound to an object (like an analysis)
that is in another process, that process will know nothing about the listener
id when the bindlistener call hits it.  That call hits the process because
we have to proxy it to the process.  And as usual, we had to do this because
some objects like ptrace targets require the RPC logic to run in the thread
that is monitoring the ptrace'd target.

Nothing else to do but sigh.