Prototype of the Analysis server: i.e. can run ROP and report results.
Improve the monitor callback interface slightly: the char * buf that the
callback receives is now NULL-terminated, even though it might not
technically be a string.  The int len arg does NOT cover that NULL char!

Make rop_checkret capable of outputting in the SimpleResult format
scanned by the XML SOAP analysis server.