Smarter probing-disasm linkage; supports caller custom disasm handling.
Needed a way to place probes on specific instr types, like
probe_register_function_instrs already does; BUT also needed to only do
it for some of them, based on the disasm info.  So now there is a callback
that influences that.

This helps us listen on jumps, but not to listen on jumps that are
function-internal, or are absolute jumps.  Need this to follow jmps
in CFI.