In CFI, handle RETs with immediates specially.
I had been assuming all RETs to be equal, but this is not the case, as ld.so
revealed to me.  I found that on (at least) one 32-bit machine, the code to
resolve a dynamic symbol would use RET with immediates to effect a "return to
lib X" control flow change.  RET with an immediate is different than a normal
RET pattern, since in a normal pattern, you clear out the stack frame all the
up to the place the retaddr is stored on the stack, then RET.  With an immed
value, RET pops the retaddr, *then* releases stack space by the amount
specified in the immed value.  So what it really does is returns to a new
addr, then cleans out the current frame on the stack.  Perfect for dynamic
symbol resolution; but I didn't anticipate it.

So, when we disasm, we notice these RETs, and if we hit one and the retaddr
is not what it should have been, we instrument the "return" address instead
of flagging a violation.

This is starting to look pretty good.