Improve CFI's ability to get through big, real apps.
Basically, this commit is a bunch of changes that help CFI be smarter
about what it does when control flow branches to an address that is not
part of some symbol (and thus we can't safely disassemble the region
that contains it).  Eventually, we will just single step until we hit a
known region; but not yet.

So, for now, we try even harder not keep "untracked" areas in our shadow
stack.  One part is not done, but this is a big improvement...