Implement most of the rest of analysis RPC interface.
The big thing missing is UploadAnalysis, which has questionable value
at this point.

Also, the service is only implemented for analysis binaries that don't
provide "external control" -- meaning they do not communicate with the
server to receive proxied requests; we don't need this yet; and at best
it is a nice feature allowing the user to apply the target API to
targets within an analysis, etc.