Don't load the current thread from its kernel stack.
(Obviously bogus; the current thread CPU state is in the CPU, not on the
stack!  This probably waited until now to manifest because we now
nuke the debug register contents when we open a Xen target, so the
current thread is always marked dirty immediately, so it gets flushed.
Then when the user (via dumpthreads) called target_load_available_threads,
the current thread CPU state was being wrongly loaded from the task_struct
cached CPU state in mem on the stack -- even if it was the currently running
thread.)