Automatically resolve symbol declarations for which a definition is available.
The Linux kernel uses decls of struct types all over the place, without
defining them, for some object files.  So to avoid ickiness of having an
instance symbol which only has a declared type, and needing to resolve the
declaration at instance symbol load time in the target, reuse defined
symbols and types for the declared symbols, when they are available.