Add an optional root_prefix path to debug/binfiles.
This is an alternate prefix that will be prepended to any files
that the binfile code attempts to open.  This helps us load
binary file information for filenames whose binaries are not
actually at that location in the / filesystem that the library
code is running on.  For instance, this is useful when we use the
target library to open ELF/debuginfo for files that are really
inside a VM that we are inspecting; it helps us look for them in
this prefix instead of our root.

Why put this here?  Because the library attempts to infer new
files to open sometimes (i.e. as it loads debuginfo), so we have
to make sure it looks in the right place.  When the user says to
open a binary file as a debugfile, they must use the real path
(i.e., including @root_prefix below).  But, those files
themselves may have embedded links to other files -- and if those
links are absolute, we have to prepend our @root_prefix.