Add overlay support to target backend API; add process overlay for xen.
The idea is that underlying targets must provide a list of their thread
ids that are available to be instantiated as overlay targets; a user can
then call target_instantiate_overlay on one or more of those thread IDs.
There are backend functions to lookup thread IDs by name or id.

The overlay target itself must provide a mechanism for handling debug
events detected by the underlying target.

The xen-process overlay target adds layers atop Xen that process even
more of a task's PCB (its mmaps, primarily), and load all the debuginfo
for those files.  It handles simple breakpoints and single stepping.

The first version does not track changes to the mmap or to threads in
the thread group.  Also, it will not instantiate the overlay as expected
if that task/thread id passed in is not a group leader; it will find the
group leader and use *that* as the overlay thread.