Fix current_thread_ptr for x86_64.
We always have to be able to figure out which kernel thread is running
so we can figure out which task is running.  This is easy on i386; harder
on x86_64.

For i386/x86:

 The bottom of each kernel stack has the thread_info struct; the first
 pointer in the thread info struct is to the task_struct associated
 with the thread_info (i.e., thread_info->task).  So, if we want
 thread_info, just load the value at current_thread_ptr; if we want
 the current task_struct, load the first pointer at
 current_thread_ptr, and deref it to load the current task_struct's
 value.

 For x86_64:

 kernel_stacks are always per_cpu unsigned long "pointers", even if
 there is only one CPU.  The value of
 xstate->kernel_stack_percpu_offset is an offset from the kernel's
 %gs.  So we have to grab the saved %gs (which Xen places in
 target->global_thread->state->context.gs_base_kernel), then apply the
 offset, then we have our pointer.

 (This works for 3.8.4; we'll see how much/often the x86_64 percpu
  stuff has/will change over time.)

It is non-obvious how to handle percpu data in general; the debuginfo
is a lie -- the user has to know the variable is percpu to do anything
useful.