Infer single steps if contents of debugreg6 are not what we expect.
(Same fix as for xen target; dr6 does not have the right bits set.  BUT,
when I was looking in the hypervisor src yesterday, I discovered that
Xen saves the interrupt vector info in the guest context struct, it
looks like -- so we can get rid of some uncertainty about this in the
debug event handling code -- but still have to distinguish between
hw breakpoints and single steps).