Support mmap dynamism for the xen-process target.
We can now track (intelligently) changes to a process's mmap, and
load/unload debugfiles for the new/stale regions and ranges.  It is
passive initially -- we reexamine the mmap every debug event -- but
we do our best to cache values and load only when necessary.  Active
probing would be done using probes on mmap, mprotect, etc.

One word of warning: if the user has symbols or values that are bound
to a region that gets deleted, they will fail.  The solution to this
of course is to keep a "hashed" bit in the region struct, and warn the
user if they try to access a value/symbol that is bound to "unhashed"
region.  I'll do this eventually.