Initial round of Xen HVM support (registers, sw breakpoints).
There were two primary changes to get HVM software breakpoints
going.  First, I had to use the custom hvm (get|set)context
functions, rather than the vcpu getters/setters, to retrieve
current CPU state.  They behave very similarly, but it seems
that only the HVM functions will properly retrieve stuff like
the segment registers, LDT, some stuff from the VMCS, etc.  So
there are wrappers around CPU context get/set, now.  Some of the
register stuff is still missing; I'll fill it in later.

The second change: VMX provides a Monitor Trap Flag to allow
the external (to domU) debugger a way to singlestep
the guest without reusing the guest's CPU's EFLAGS (EF_TF).
This has the nice property that it is invisible to the guest...
but it also means that stepping will continue relative to the VCPU,
so the debugger can step nicely across thread context switches.  This
is unlike debugging a PV guest which will only singlestep threads
-- single step will not persist across a context switch, because
EFLAGS will be pushed and the new thread's EFLAGS popped.

What is really not cool is that the Xen hypervisor basically forces
the debugger to use the MTF if the CPU supports it.  If it exists
and the debugger doesn't use it, then the debugger is still notified,
but so is the guest -- and there isn't a way to block the notification
to the guest.

Of course, there isn't really an analogue to the MTF for hw debug
registers, and certainly no way to do sw breakpoints like that.  So the
debugger still has to use guest hw debug resources when needed.

The bad news: there seem to be at least a couple different bugs with
Xen's HVM implementation in how it handles debug registers -- i.e.,
you don't really see updates to DR6 correctly, unless you enable at
least one hw debug register via DR7.

And then, hw debug exceptions are not reported to the debugger -- BUT
they are correctly reported via DR6 at the next debug exception!

So, this stuff will have to be fixed before I can make hw debug
support happen.