__xen_vm_load_current_thread must clear errno if it returns global thread.
Even if there was an error loading the current thread, if this function
tries to return the global thread instead, it is doing best-effort and thus
must clear errno.  The reason is that xen_vm_(read|write)_reg depend on
errno to detect errors, *but* calls to them might result in calls to this
function -- and thus we don't want to falsely propagate errno.