Add support for shared page breakpoints in Xen process target.
This required several things.

  1) Target backend API addition for translating virtual target addrs
     to physical ones.
  2) Target backend API additions to read/write physical addrs.
  2) Adjustments to memmod abstraction (memmods are either physical
     or virtual address-based).
  3) Improved pgd/cr3 loading.

Overall, this is a twisted feature.  For history... we needed shared
page breakpoints in the xen-process target because if we place a sw
breakpoint in a page of a process in a VM, that page might be (is very
likely to be) shared with other processes.  That means, even if we
emplaced the sw breakpoint on behalf of pid X, if pid Y has that same
physical page mapped into its virtual addrspace, pid Y will see a debug
exception.  Of course, there is no overlay xen-process target to handle
it; so we must handle it in the underlying xen target.  Memmods help us
do that.

Basically, the xen-process target only ever inserts sw breakpoints at
physical addrs; it assumes that any page it wants to emplace a breakpoint
in could be shared (it's likely to be true anyway and is much easier than
trying to actually track the sharing status of physical pages).  So it
uses a target helper function to insert a breakpoint at a physical address.
The xen target, if it can't find a breakpoint at a vaddr when a breakpoint
debug event happens, translates it to a paddr and checks again.  If there
is an memmod at that paddr, but no overlay target to handle it, then we
emulate the breakpoint.  We don't construct an artifical probepoint, like
normal; that would be icky because probepoints, probes, threads are strongly
bound to a specific target, and we use that fact repeatedly throughout the
code.  Thus, it is impossible to share a probepoint between targets.  This
is why probepoints now reference memmods -- which are bound to a target --
BUT the memmod may be bound to a different target than the probepoint it is
associated with.

All this ickiness comes from the fact that we assumed that a single target
has a single address space, way back when.  We assumed that threads could
share that address space.

Now we have to support the case where there is a de facto underlying
addrspace (the VM's physical space; its entire memory) that can be shared
by all threads.  Our concept of address spaces simply did not support this,
and it was impossible to try to express the fact that a single thread
really has two address spaces.

So I broadened the target address abstraction slightly.  Now target
addrs are by default *virtual* addrs; if the target supports it, they
may be translated to an underlying physical address.  These physical
addresses are not formally bound to an address space; they are only
useful for target_(read|write)_physaddr.

If we ever hit this problem in another target, the solution would be to
formalize this a bit more by having hierarchies of address spaces, with
a more formal translation API.  But on this project, we'll never hit that.

This is unpleasant, and the sheer amount of cases/hypervisor bugs/"situations"
we now support in the Xen target is pretty insane.

A final note: this commit only works with libvmi.  I have also only tested
it with HVM and Xen 4.x