Warn on something that would signal a race in the underlying target.
If the target doesn't provide individual thread control, we might run
into this, so VERROR on it so the user notices.

In general, there is no good solution to these problems.  Hard to debug
a multithreaded target that doesn't let us control threads individually
(either by pausing all other threads while we alter memory on behalf of
one thread; or by providing a per-thread copy-on-write memory).

There are always solutions (like making pages read-only; catching thread
context switches and altering memory modifications that might affect the
thread so that the memory is consistent with what the incoming thread
expects; or *temporarily* giving the target thread's addrspace a
copy of the page we're editing; etc), but they tend to be heavyweight,
causing extra traps.

Anyway, I'm writing this all down here because I'm thinking about it, and
because we need to abstract all this messy thread control behavior out so we
can control the crazy cases better.  Each target could conceivably provide
different kinds of abstractions to control threads, and provide the various
thread control op modes (STRICT, LOOSE) -- we need to wrap this stuff all
up so that all the special case code scattered all over is easier to
understand and reuse.