Add support for handling "interrupted" breakpoint single stepping.
On some targets (xen-process overlay ones, for now), once we setup a
singlestep after hitting a breakpoint in userspace, we may find that
that thread is single stepping correctly, but in kernel space!  This
can happen if we have to use HVM global monitor trap flag instead of
EFLAGS TF.  Even if we had setup the MTF to single step the guest in
userspace, that may not be what happens.  For instance, suppose the
instruction causes a page fault, or that a clock interrupt happened.
We'll find ourselves stepping in the kernel, in its handlers, I believe.
We can assume the thread did *not* do its singlestep of the breakpoint's
original instruction.  If it had, the EIP for the MTF event would still
be in userspace -- because single step debug exceptions are traps
following an instruction's execution.  Thus, we need to put the breakpoint
back into place and remove all state setup to handle it, EXCEPT to note
down that this thread's overlay SS was interrupted at probepoint X, but
that the prehandler was already run.  This way, we won't run the
prehandler again.

(This of course is somewhat bogus, because it might affect vCPU state (we
hit the BP twice instead of just once)... but whatever.  There's nothing
else we can do if we are forced to use the Monitor Trap Flag, like HVM
on vanilla Xen 4 forces us to do.)

So, I exposed this as a "general" mechanism.  There is now an
"interrupted_ss_handler" that handles the case where a thread was
interrupted before it could single step.  It does *not* handle the case
(yet) where it had done at least one single step and was continuing
stepping past the breakpoint.  Don't try that at home!!

Of course, it's only hooked in for this specific xen-process case.

Docs for probepoint_interrupted_ss_handler:

* This function handles the case where a thread is singlestepping, but
* it has either stepped into a new context (i.e., user into kernel), or
* into a different thread.  Initially, we only handle the case where
* the breakpoint was just hit and the single step setup, but not run.
* We do not handle the case where a single step action was running.
*
* This is much the same as the single step handler, BUT we abort any
* running actions; we do not continue single stepping; and we unwind
* all state associated with handling the breakpoint.
*
* NB: we cannot read any register state from the thread; another thread
* or thread context (i.e., kernel instead of user) might have been
* entered.  We just want to unwind the previous context and fixup the
* probepoint.