Add overlay support to cfi_check.
If you specify an overlay, the symbols you pass on the command line
(the root symbols in which initial CFI probes are placed) apply only
to the overlay target, of course.

(This stuff needs to be factored out, somehow, because there's lots of
shared code to instantiate an overlay target, but it's not clear quite
how to do it because overlays are pretty custom to the application
using them.  Well, ok, in 10 seconds of thinking it's more clear, but
this works for now.)