Fix race in xen_vm_pause .
Sometimes we can get into the situation where the target status thinks it's
running, but dominfo says it's paused.

We could perhaps choose to handle this by changing target state not only
on command, but every time we load dominfo.  That's not how it's currently
written though, so just handle this...