Add a model for target-specific "personalities".
Ok, in the source, they're just called "kinds".  The idea is that a
target can be either an OS or process instance (in the future we'll
add an application instance too).  Kinds are supposed to provide
operations that wrap common abstractions that target kinds provide.
For instance, all OSes provide system calls, signals, processes,
memory mapping, etc.

OSes current expose version info and syscalls.  Processes are just stubbed
for now until what we need becomes clear, or I have spare time.

This is all pretty straightforward, except that for OSes, I opted to
have some calls that build and register syscall probes directly, instead of
implementing a separate, OS/target-independent probe library to do it.
When you get down to it, probing the syscall return path is complicated, and
is a function of both the arch (which I haven't factored out yet) and OS.

Obviously the only OS I implemented is Linux, and the per-syscall probe op
is not implemented yet.  Also, we assume any HVM guest is Linux :).