Add probe values and filters.
Biggest change: probe_handler_t has changed to accomodate more info.  Most
important is that the tid_t that the event happened in is passed; it was
a bad oversight to leave this off when I added thread support.  Also, now
I pass the probe itself, the triggering probe (the one sitting directly
underneath the probe itself (either a base probe or a source probe), and
the base probe (i.e., if the triggering probe is a source probe, base
will be the probe registered atop the lowest-layer probepoint).

The idea with probe values is that some probes naturally provide a set of
values.  Furthermore, it is convenient to filter on these values.  Examples
of probes that provide values are watchpoints and function entry/exit probes.

Then once you have this means of autoloading values before/while pre/post
handlers are executed, it becomes natural to attach filtering probes as
sinks of value probe sources -- so that the pre/post handlers of the filters
are only called when their filter matches.

This way, the filtering mechanism is abstracted from the specific probe.  It
just requires that the probe under the filter can load the values it is trying
to filter on.

Right now, there is no way to load "default" values for a probe; specific
probe_ops must be built to enable this for probe types.  I added two new
probe types that take advantage of it: function entry(pre)/exit(post) probes
(well, they're not new, but I wrapped the existing code to have value support);
and watchpoints (pre=previous value; post=new value).

Right now, the only kind of filter is an ANDed list of regexps.  Good enough
for now.

This was kind of an experiment.  We need a way to do this for real, but I
can't have more hacks like vmprobegeneric... so at least now we have a
regular mechanism that makes some sense.

One thing I didn't add, but that I will, is context filtering (context
filtering is filtering the probe pre/post handlers based on thread context
too).  Different targets will provide different context information.