Make sure we catch and handle "asynch", pending debug exceptions.
Since we allow pauses to be commanded asynchronously
w.r.t. target vm execution state, we have to check if there is
something to handle once we successfully pause it, and handle it
if so.  Otherwise if a target_pause() and debug exception happen
at the "same" time relative to the user, we might leave a debug
event unhandled, and this could whack the target.