Add --use-os-syscall-probes to enable special syscall probes.
This option should only be used if you're probing all/most syscalls;
for Linux, instead of placing probes on the entrance/exit of syscall,
it only places one on the entry, and then a single probe on the SYSRET
to userspace.

There might be some syscalls that return in weird ways that you would
rather catch via the global ret to userspace path too.

Anyway, this behavior is no longer the default; it adds too much overhead
for probing small sets of syscalls.

Also fixup some argp handling int key stuff.