Abandon trying to read copy_process.p in thread_entry active probe.
Just read %eax and "cast" it instead.  The DWARF info seems to lie; p
is technically available at RETQ but it depends on the frame_base value
which seems to be incorrect -- the stack has already been cleaned up but
the frame base values don't seem to account for that.  Oh well.

Also fixup some more probe RESULT_ERROR cases.