Add untested probe_register_function_invocations .
This is designed to register probes on all calls of function Y inside
function X; it places a probe on the call, then on the instruction after
the call.  Then the caller's probe is attached to all those probes; its pre
handlers are called when the calls are hit; its post handlers are called
after the call has completed.  If the caller's probe doesn't have pre/post
handlers, probes are not registered on the call or post-call instruction.
It also handles inline instances of Y inside X by placing
a probe on the first instr of the inline instance; then on the first instr
following the inline instance.  If the inline instance is non-contiguous,
we just take its lowest and highest addresses and hope for the best.  No
sane optimizer could possibly overlap two inline instances of the same
origin.  Oh well, they can do anything.

It's untested because I thought I needed it, but now I don't.  So just
leave the code around for now... we'll want it someday.