Extend function entry/exit probes to deal with external absolute JMPs.
If @follow_jumps is set, we will look for JMPs that leave the body
of the given function.  We will then place probes on RETs in the target
function (or safe disasm segment) containing the JMP target.  This is
done recursively to achieve "total" coverage.

Very much like how we handled CFI.