Support newer kernel module layouts too.
This is now bad -- the section layout info is in /sys/module...; but all
the code was already written to emulate the module loader.  So keep using
that for now, and hack it to support newer layout conditions.

The biggest change is to support aligning major chunks (i.e., text, text+ro,
init) onto page boundaries.  Also ensure we check kallsyms (to map symtab
and strtab if CONFIG_KALLSYMS) appropriately.  Also support newer layouts
where symtab and strtab are placed after the init section allocations.

(Still buggy for newer kernels; .init and .symtab/.strtab sections are wrong;
but this does not bother VMI really.)