Finally!  Add a DWARF expression evaluator (with location desc support).
We have waited over a year for this (or is it more?).

Now we attempt to evaluate complex location expressions when loading
values or computing their addresses.  Unfortunately, I had decided to
use two or three outvars for most of my location resolution API
(addr or reg or offset); this choice is now undone and we actually
use a single struct location outvar throughout most of it.  There are
simply too many possible "locations" -- sometimes even they are not
locations, but rather implicitly computed values taking the place of a
location (i.e., when a value is no longer in a register or in memory,
but the compiler could still compute its value).

Unfortunately, gcc has increased its powers of DWARF emission... so
supporting the easy expressions will not do.  Moreover, there is now a
wonderful DWARF_OP_call_frame_cfa that can be used in location
expressions -- but is a DWARF expr call() to get the CFA for the frame
in question.  And of course now gcc uses this liberally as the
pseudo frame_base register (which itself is used heavily to compute
local var addresses in functions).  So now I *have* to write a CFA
interpreter too... oh well, that will lead to backtrace support!

But, for older gccs or if you make sure to compile only DWARF v2
debuginfo, this might be ok.

The major ops not supported are

  DW_OP_xderef, DW_OP_xderef_size
    (these support dereferencing pointers in different addrspaces; I
     highly doubt we'll ever encounter a target with an ABI that actually
     supports such a feature!)
  DW_OP_push_object_address
  DW_OP_tls_address
  DW_OP_call_frame_CFA
  DW_OP_call{2,4,_ref}
    (we don't support calls to other DWARF expressions!)
  DW_OP_piece, DW_OP_bit_piece
    (we do not support composite locations/values, where the location or
     value described might be sprinkled all over the place -- ugh!  This
     is unfortunately somewhat important; gcc does emit this junk.)

And we don't support any GNU ops, like

  DW_OP_GNU_push_tls_address, DW_OP_GNU_uninit,
  DW_OP_GNU_encoded_addr
    (haven't even looked at these)
  DW_OP_GNU_implicit_pointer
    (heavily used on current gccs)

nor more recently and definitely used

  DW_OP_GNU_entry_value, DW_OP_GNU_regval_type

So, I suppose the "future" priority list is

  DW_OP_piece, DW_OP_bit_piece
  DW_OP_GNU_implicit_pointer
  DW_OP_GNU_entry_value (but this might need DW_TAG_call_site
    support to be really useful)
  DW_OP_call{2,4,_ref}
  DW_OP_call_frame_CFA

and I can't imagine ever implementing the rest.