Inaugurate debugfile_ops/debugfile_type_t and add a DWARF CFA parser.
I started factoring out per-debugfile-type ops, and used that to add
a DWARF CFA parser (well, it actually parses real DWARF CFA< and the
slightly augmented format generated by gcc).

The strategy is to scan the whole .(debug|eh)_frame section, parsing all
CIEs and FDEs, but not evaluating the FDE programs until we need them (but
CIE initial programs (which set default register rules) are run immediately).
This tradeoff seems to be a good one.  I could have also used .eh_frame_hdr
to save more memory, but it's not worth it even for large debuginfo files,
and I didn't realize it in time.

Now to figure out how to leverage this to build struct location_ctxt to
allow per-frame symbol resolution and stack unwinding... but this is the
hard part.