Really fix the "missing CFI section contents" problem.
I hadn't noticed that the section was NOBITS -- obviously the CFI info
has to go in the primary executable/library, not in the debuginfo file
-- so now we try to load "missing" sections from the "pointing" binfile
(i.e., the binfile for the executable/library, not the one containing
debuginfo) if the debuginfo binfile doesn't have the content.

This pretty much works, except that on some platforms, the addresses in
the ELF executable/library binfile are absolute/relative and the ELF
debuginfo file is relative/absolute (i.e., not the same).  This causes
a bunch of heartburn, and it's not really fixed yet...