Add support for passing inputParams to analyses via Instantiate().
We already had a bunch of this in place, but I wasn't doing anything
with it.  Now you can pass, via AnalysisSpecT, some inputParams (each of
type NameValueT).  The named params can match a listed input param in the
analysis description, OR NOT.  If the given params don't match a param in
the analysis description, then we add them as switched params to the
command line that will fork the analysis (if the NameValueT name is
strlen(name) > 1, we do --<name>; else we do -<name>).  This is handy for
passing arbitrary params to the underlying binary, even if they are not
listed in the analysis description.xml file.

When input param names *do* match an existing param in the description
(and they may match either the long name or the regular name), we handle
them as specified in the description.  If the description said they were
command line params (via the isCommandLine ParamT attribute), we pass them
on the command line (otherwise for now we ignore them, module the above
"unspecified" case).  If the isCommandLineSwitched param is set, we pass
them as switched params (i.e., --<longName <value> or -name <value>).  If
they had a longName in the description, we always use that; otherwise we
use the short name.

Note to people writing analysis descriptions: you had better make sure
that if your ParamT isCommandLineSwitched, that its name is a single letter,
or that the program you are invoking really takes -FooBarBaz params.

Finally, analysis_spf_xen.py.in includes examples of this stuff.